+++ mac-installer.sh
@@ -20,3 +20,12 @@ heading () {
 title
 
 # - - - - - - - - - - - - - - - - -
+
+if [ -d "/Library/Developer" ]; then
+  printf "\n$(tput setaf 2)# Xcode Command Line Tools already installed.$(tput sgr0)\n"
+else
+  heading "Installing Xcode Command Line Tools ..."
+  xcode-select --install
+fi
+
+# - - - - - - - - - - - - - - - - -